unit tests for common code
