+++ .github/workflows/run_unittests_common.yml
@@ -3,33 +3,13 @@ on:
   schedule:
     - cron: '0 1 * * *'
 
-  workflow_dispatch:
-    inputs:
-      CUR_TIMESTAMP:
-        description: Timestamp for the build
-        required: false
-        # Leave empty; we'll set it programmatically if not provided
-        default: ''
-
 jobs:
-  build-images:
+  run_tests:
     runs-on: ubuntu-24.04
     steps:
 
       - name: Checkout code
         uses: actions/checkout@v4
-
-      - name: Set Timestamp
-        id: set-timestamp
-        run: |
-          if [[ "${{ github.event.inputs.timestamp }}" == "" ]]; then
-            #echo "timestamp=$(date +%Y-%m-%dT%H:%M:%SZ)" >> $GITHUB_ENV
-            echo "timestamp=$(date +%Y%m%d_%H%M)" >> $GITHUB_ENV
-          else
-            echo "timestamp=${{ github.event.inputs.timestamp }}" >> $GITHUB_ENV
-          fi
-        env:
-          TIMESTAMP: ${{ github.event.inputs.timestamp }}
 
       - name: Set up QEMU
         uses: docker/setup-qemu-action@v3
