+++ src/common_unittests/common_unittests.vcxproj
@@ -19,7 +19,7 @@
     </ProjectConfiguration>
   </ItemGroup>
   <PropertyGroup Label="Globals">
-    <ProjectGuid>{62926c5e-ad62-4385-abdb-cc2c70147357}</ProjectGuid>
+    <ProjectGuid>{73d20930-47ca-4439-bb62-294da213de01}</ProjectGuid>
     <Keyword>Win32Proj</Keyword>
     <WindowsTargetPlatformVersion>10.0.22621.0</WindowsTargetPlatformVersion>
     <ConfigurationType>Application</ConfigurationType>
@@ -99,15 +99,33 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
+    <ClCompile Include="..\common\ConfigManager.cpp" />
+    <ClCompile Include="..\common\ConfigSetup.cpp" />
+    <ClCompile Include="..\common\ConfigSetupItem.cpp" />
+    <ClCompile Include="..\common\Platform.cpp" />
+    <ClCompile Include="ConfigManagerTests.cpp" />
     <ClCompile Include="main.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\duplitrace_indexer\duplitrace_indexer.vcxproj">
-      <Project>{23eecbe0-79c1-4eb9-b2f0-e9b4102e02f7}</Project>
-    </ProjectReference>
+    <None Include="packages.config" />
+    <CopyFileToFolders Include="test_configs\valid_config.cfg">
+      <DeploymentContent Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">true</DeploymentContent>
+      <FileType>Document</FileType>
+    </CopyFileToFolders>
+    <CopyFileToFolders Include="test_configs\invalid_config_non_int.cfg">
+      <DeploymentContent Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">true</DeploymentContent>
+      <FileType>Document</FileType>
+    </CopyFileToFolders>
+    <CopyFileToFolders Include="test_configs\invalid_config_wrong_str_value.cfg">
+      <DeploymentContent Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">true</DeploymentContent>
+      <FileType>Document</FileType>
+    </CopyFileToFolders>
   </ItemGroup>
   <ItemGroup>
-    <None Include="packages.config" />
+    <ClInclude Include="..\common\ConfigManager.h" />
+    <ClInclude Include="..\common\ConfigSetup.h" />
+    <ClInclude Include="..\common\ConfigSetupItem.h" />
+    <ClInclude Include="..\common\Platform.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
