+++ src/common_unittests/common_unittests.vcxproj.filters
@@ -1,0 +1,54 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+  <ItemGroup>
+    <ClCompile Include="ConfigManagerTests.cpp" />
+    <ClCompile Include="main.cpp" />
+    <ClCompile Include="..\common\ConfigManager.cpp">
+      <Filter>indexer_src</Filter>
+    </ClCompile>
+    <ClCompile Include="..\common\ConfigSetup.cpp">
+      <Filter>indexer_src</Filter>
+    </ClCompile>
+    <ClCompile Include="..\common\ConfigSetupItem.cpp">
+      <Filter>indexer_src</Filter>
+    </ClCompile>
+    <ClCompile Include="..\common\Platform.cpp">
+      <Filter>indexer_src</Filter>
+    </ClCompile>
+  </ItemGroup>
+  <ItemGroup>
+    <Filter Include="test_config_files">
+      <UniqueIdentifier>{d79e5ff7-d93b-43de-b120-4cecf18927d6}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="indexer_src">
+      <UniqueIdentifier>{9838e861-a078-4ce9-99fd-e45001416736}</UniqueIdentifier>
+    </Filter>
+  </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="..\common\ConfigManager.h">
+      <Filter>indexer_src</Filter>
+    </ClInclude>
+    <ClInclude Include="..\common\ConfigSetup.h">
+      <Filter>indexer_src</Filter>
+    </ClInclude>
+    <ClInclude Include="..\common\ConfigSetupItem.h">
+      <Filter>indexer_src</Filter>
+    </ClInclude>
+    <ClInclude Include="..\common\Platform.h">
+      <Filter>indexer_src</Filter>
+    </ClInclude>
+  </ItemGroup>
+  <ItemGroup>
+    <CopyFileToFolders Include="test_configs\valid_config.cfg">
+      <Filter>test_config_files</Filter>
+    </CopyFileToFolders>
+    <CopyFileToFolders Include="test_configs\invalid_config_non_int.cfg">
+      <Filter>test_config_files</Filter>
+    </CopyFileToFolders>
+  </ItemGroup>
+  <ItemGroup>
+    <None Include="test_configs\invalid_config_wrong_str_value.cfg">
+      <Filter>test_config_files</Filter>
+    </None>
+  </ItemGroup>
+</Project>
