+++ src/common_unittests/main.cpp
@@ -1,0 +1,6 @@
+#include <gtest/gtest.h>
+
+int main(int argc, char** argv) {
+    ::testing::InitGoogleTest (&argc, argv);
+    return RUN_ALL_TESTS ();
+}
