+++ src/common_unittests/test_configs/invalid_config_non_int.cfg
@@ -1,0 +1,4 @@
+[test_section]
+str_value_with_valid_values=value_2
+str_value=ValidStr
+int_value=NotAnInt
