+++ src/common_unittests/test_configs/invalid_config_wrong_str_value.cfg
@@ -1,0 +1,4 @@
+[test_section]
+str_value_with_valid_values=value_20
+str_value=ValidStr
+int_value=1701
