+++ src/duplitrace.sln
@@ -7,6 +7,8 @@ Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "duplitrace_indexer", "dupli
 EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "indexer_unittests", "indexer_unittests\indexer_unittests.vcxproj", "{62926C5E-AD62-4385-ABDB-CC2C70147357}"
 EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "common_unittests", "common_unittests\common_unittests.vcxproj", "{73D20930-47CA-4439-BB62-294DA213DE01}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|x64 = Debug|x64
@@ -31,6 +33,14 @@ Global
 		{62926C5E-AD62-4385-ABDB-CC2C70147357}.Release|x64.Build.0 = Release|x64
 		{62926C5E-AD62-4385-ABDB-CC2C70147357}.Release|x86.ActiveCfg = Release|Win32
 		{62926C5E-AD62-4385-ABDB-CC2C70147357}.Release|x86.Build.0 = Release|Win32
+		{73D20930-47CA-4439-BB62-294DA213DE01}.Debug|x64.ActiveCfg = Debug|x64
+		{73D20930-47CA-4439-BB62-294DA213DE01}.Debug|x64.Build.0 = Debug|x64
+		{73D20930-47CA-4439-BB62-294DA213DE01}.Debug|x86.ActiveCfg = Debug|Win32
+		{73D20930-47CA-4439-BB62-294DA213DE01}.Debug|x86.Build.0 = Debug|Win32
+		{73D20930-47CA-4439-BB62-294DA213DE01}.Release|x64.ActiveCfg = Release|x64
+		{73D20930-47CA-4439-BB62-294DA213DE01}.Release|x64.Build.0 = Release|x64
+		{73D20930-47CA-4439-BB62-294DA213DE01}.Release|x86.ActiveCfg = Release|Win32
+		{73D20930-47CA-4439-BB62-294DA213DE01}.Release|x86.Build.0 = Release|Win32
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
