+++ src/indexer_unittests/indexer_unittests.vcxproj.filters
@@ -1,45 +1,9 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <ItemGroup>
-    <ClCompile Include="ConfigManagerTests.cpp" />
-    <ClCompile Include="..\common\ConfigManager.cpp">
-      <Filter>indexer_src</Filter>
-    </ClCompile>
-    <ClCompile Include="..\common\ConfigSetup.cpp">
-      <Filter>indexer_src</Filter>
-    </ClCompile>
-    <ClCompile Include="..\common\ConfigSetupItem.cpp">
-      <Filter>indexer_src</Filter>
-    </ClCompile>
-    <ClCompile Include="..\common\Platform.cpp">
-      <Filter>indexer_src</Filter>
-    </ClCompile>
+    <ClCompile Include="main.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <Filter Include="configs">
-      <UniqueIdentifier>{7e6076bd-e199-4bcc-8b07-8eab90255a91}</UniqueIdentifier>
-    </Filter>
-    <Filter Include="indexer_src">
-      <UniqueIdentifier>{1e159805-2101-4a7f-9021-1c4974fb9e26}</UniqueIdentifier>
-    </Filter>
-  </ItemGroup>
-  <ItemGroup>
-    <CopyFileToFolders Include="test_configs\valid_config.cfg">
-      <Filter>configs</Filter>
-    </CopyFileToFolders>
-  </ItemGroup>
-  <ItemGroup>
-    <ClInclude Include="..\common\ConfigManager.h">
-      <Filter>indexer_src</Filter>
-    </ClInclude>
-    <ClInclude Include="..\common\ConfigSetup.h">
-      <Filter>indexer_src</Filter>
-    </ClInclude>
-    <ClInclude Include="..\common\ConfigSetupItem.h">
-      <Filter>indexer_src</Filter>
-    </ClInclude>
-    <ClInclude Include="..\common\Platform.h">
-      <Filter>indexer_src</Filter>
-    </ClInclude>
+    <None Include="packages.config" />
   </ItemGroup>
 </Project>
